refactor: separate star generation
